requirements/prod.txt
-r ../vendor/src/funfactory/funfactory/requirements/prod.txt
django_browserid
requests
